design finally decent, moved utils out, filtering and sort good
