+++ frontend/src/lib/logout.js
